final_requirements.txt
# RAG 시스템 핵심 라이브러리

langchain-core==0.3.74

langchain-community==0.3.27

langchain-google-genai==2.1.9

langchain-openai==0.3.29

langchain-text-splitters==0.3.9

chromadb==1.0.20

sentence-transformers==5.1.0

flashrank==0.2.10

onnxruntime==1.22.1



# 웹 서버 및 API

fastapi==0.116.1

uvicorn[standard]==0.35.0

pydantic==2.11.7

python-dotenv==1.1.1



# 데이터 처리 및 유틸리티

pandas==2.3.2

numpy==2.2.6

tqdm==4.67.1

typing-extensions==4.14.1

torch==2.8.0

kiwipiepy==0.21.0

kiwipiepy-model==0.21.0

langgraph==0.6.6
